shop_service/pyproject.toml
[project]
name = "shop-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Manh Pham",email = "manhpham90vn@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "motor (>=3.6.1,<4.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dependency-injector (>=4.45.0,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
